+++ certbot-apache/certbot_apache/configurator.py
@@ -32,7 +32,7 @@ from certbot.plugins.enhancements import AutoHSTSEnhancement
 from certbot_apache._internal import apache_util
 from certbot_apache import constants
 from certbot_apache._internal import display_ops
-from certbot_apache import http_01
+from certbot_apache._internal import http_01
 from certbot_apache._internal import obj
 from certbot_apache._internal import parser
